Add bounds checking for entity movement, now should stay on screen if it tries to go off of it
- change position change args to int16, didnt want to deal with unsigned wrapping
- added constexpr functions in utility.h for min, max, clamp
- entity now wraps new incoming positions to [1, screenwidth - ent width - 2] for safety the two is needed due to inclusive/exclusive mixing

+++ Starship_GalaxyWars/include/Entity.h
@@ -24,7 +24,7 @@ public:
 	uint16_t* GetTextureData();
 	uint16_t GetTextureDataSize();
 
-	void SetPosition(uint8_t x, uint8_t y);
+	void SetPosition(int16_t x, int16_t y);
 	void SetPosition(XYCoord newPosition);
 	void SetLastRenderedPosition(ScreenRegion region);
 	void SetRenderDirty(bool dirty);
